// server/src/api/v1/interfaces/SlipData.ts
interface SlipData {
    _id: string;
    spAS: number;
    cId: string;
    cN: string;
    gN: string;
    dT: string;
    cc: number;
    uId: string;
    uP: string;
    uN: string;
    gId: string;
    st: string;
    po: string;
    rk: number;
    eF: number;
    mS: number;
    cT: number;
}

export { SlipData };